component with person details refactored
