+++ src/App.js
@@ -1,5 +1,5 @@
 import { HashRouter, Route, Redirect, Switch } from "react-router-dom";
-import { PersonDetails } from "./layouts/PersonDetails/index";
+import PersonDetails from "./layouts/PersonDetails/index";
 import MoviesDetails from "./layouts/MoviesDetails/index";
 import { Loading } from "./layouts/Loading/index";
 import { ThemeProvider } from "styled-components";
