+++ src/layouts/PersonDetails/index.js
@@ -1,102 +1,32 @@
-import { CastContent, Content, CrewContent, List, SubdivTitle, TileContent } from "./styled";
-import { Tile } from "../Tile/index";
 import { useSelector } from "react-redux";
-import {
-  selectPeopleCredits,
-  selectPeopleDetailsStatus,
-} from "../../features/peoples/peopleDetails/peopleDetailsSlice";
-import { selectPosterSizes, selectProfileSizes } from "../../features/configurationSlice";
+
+import CreditsList from "./CreditsList";
 import Error from "../Error/index";
+import { Tile } from "../Tile/index";
+import { selectPeopleDetailsStatus } from "../../features/peoples/peopleDetails/peopleDetailsSlice";
+import { selectProfileSizes } from "../../features/configurationSlice";
 
-export const PersonDetails = () => {
-  const peopleCredits = useSelector(selectPeopleCredits);
-  const profileSizes = useSelector(selectProfileSizes);
-  const posterSizes = useSelector(selectPosterSizes);
-  const personDetailsStatus = useSelector(selectPeopleDetailsStatus);
+import { Content, TileContent } from "./styled";
 
-  return (
-    <>
-      {personDetailsStatus === "error" ? (
-        <Error />
-      ) : (
-        <>
-          <Content>
-            <TileContent>
-              <Tile personDetails profileSizes={profileSizes} />
-            </TileContent>
-            <CastContent>
-              <SubdivTitle>
-                Movies - cast({peopleCredits.cast && peopleCredits.cast.length})
-              </SubdivTitle>
-              <List>
-                {peopleCredits &&
-                  peopleCredits.cast &&
-                  peopleCredits.cast.map(
-                    ({
-                      id,
-                      poster_path,
-                      title,
-                      release_date,
-                      genre_ids,
-                      vote_average,
-                      vote_count,
-                    }) => (
-                      <Tile
-                        movie
-                        posterSizes={posterSizes}
-                        key={id}
-                        id={id}
-                        poster_path={poster_path}
-                        title={title}
-                        release_date={release_date}
-                        genre_ids={genre_ids}
-                        vote_average={vote_average}
-                        vote_count={vote_count}
-                      />
-                    )
-                  )}
-              </List>
-            </CastContent>
-            {peopleCredits.crew && peopleCredits.crew.length !== 0 ? (
-              <CrewContent>
-                <SubdivTitle>
-                  Movies - crew({peopleCredits.crew && peopleCredits.crew.length})
-                </SubdivTitle>
-                <List>
-                  {peopleCredits &&
-                    peopleCredits.crew &&
-                    peopleCredits.crew.map(
-                      ({
-                        id,
-                        poster_path,
-                        title,
-                        release_date,
-                        genre_ids,
-                        vote_average,
-                        vote_count,
-                      }) => (
-                        <Tile
-                          movie
-                          posterSizes={posterSizes}
-                          key={id}
-                          id={id}
-                          poster_path={poster_path}
-                          title={title}
-                          release_date={release_date}
-                          genre_ids={genre_ids}
-                          vote_average={vote_average}
-                          vote_count={vote_count}
-                        />
-                      )
-                    )}
-                </List>
-              </CrewContent>
-            ) : (
-              ""
-            )}
-          </Content>
-        </>
-      )}
-    </>
-  );
+const PersonDetails = () => {
+   const profileSizes = useSelector(selectProfileSizes);
+   const personDetailsStatus = useSelector(selectPeopleDetailsStatus);
+
+   return (
+      <>
+         {personDetailsStatus === "error" ? (
+            <Error />
+         ) : (
+            <Content>
+               <TileContent>
+                  <Tile personDetails profileSizes={profileSizes} />
+               </TileContent>
+               <CreditsList cast />
+               <CreditsList crew />
+            </Content>
+         )}
+      </>
+   );
 };
+
+export default PersonDetails;
